Add Retrying and pass in bot info
